backend/requirements.txt
asgiref==3.11.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
distro==1.9.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
mysql-connector==2.2.9
openai==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.1
sqlparse==0.5.4
tzdata==2025.2
pymysql           
cryptography
fastapi
uvicorn[standard]
openai
python-dotenv
requests